+++ packages/snap-preact-demo/babel.config.js
@@ -30,6 +30,6 @@ module.exports = (api) => {
 				},
 			],
 			['@babel/plugin-transform-arrow-functions'],
-		].filter(Boolean),
+		],
 	};
 };