+++ packages/snap-store-mobx/src/Search/Stores/SortingStore.ts
@@ -28,7 +28,6 @@ export class SortingStore {
 					if (!search?.query) {
 						return option.type == 'field';
 					}
-
 					return option;
 				})
 				.map((option: MetaResponseModelSortOptionMutated, index: number) => {
@@ -47,7 +46,7 @@ export class SortingStore {
 						option.default = true;
 					}
 
-					const optionObj = new Option(services, option);
+					const optionObj = new Option(services, option, index);
 
 					return optionObj;
 				});
@@ -74,14 +73,14 @@ class Option {
 	value: string;
 	url: UrlManager;
 
-	constructor(services: StoreServices, option) {
+	constructor(services: StoreServices, option, index) {
 		this.active = option.active;
 		this.default = option.default;
 		this.field = option.field;
 		this.label = option.label;
 		this.direction = option.direction;
 		this.type = option.type;
-		this.value = `${option.field}:${option.direction}`;
+		this.value = `${option.label}:${option.field}:${option.direction}:${index}`;
 
 		if (this.default) {
 			this.url = services.urlManager.remove('page').remove('sort');